Changed seeding to new file for purpose of separation of concerns
